wu-move-from-tensor-move: Move update_weights_from_tensor + helpers onto WeightUpdater (cut+paste)
Refactor chain ID: wu-move-from-tensor-move
